requirements/base.txt
yfinance
pandas
numpy
tqdm
yahoo-finance
requests
requests-cache
pandas-datareader
stockstats
Jinja2
matplotlib
mplfinance==0.12.7a10
schedule
python-dotenv
pytz
yahoo_earnings_calendar
ccxt
dataset
tweepy
peewee
dotmap
flatten-dict
python-telegram-bot
